rename: 스타일 파일들을 styled 폴더로 이동, 이름 변경, 적용
