+++ FE/issuetracker/src/components/Issue/IssueDetail.tsx
@@ -1,5 +1,5 @@
 import { useRecoilValue } from 'recoil';
-import * as S from 'components/Issue/styled.issueDetail';
+import * as S from 'components/Issue/styled/issueDetail';
 import IssueLabel from 'components/Issue/IssueLabel';
 import { issueState } from 'recoil/atoms/issue';
