+++ FE/issuetracker/src/components/Issue/index.tsx
@@ -1,4 +1,4 @@
-import * as S from 'components/Issue/styled.index';
+import * as S from 'components/Issue/styled/index';
 import IssueTitle from 'components/Issue/IssueTitle';
 import IssueDetail from 'components/Issue/IssueDetail';
 import IssueComments from 'components/Issue/IssueCommentSection';
