+++ FE/issuetracker/src/components/Issue/styled/index.js
