+++ FE/issuetracker/src/components/Issue/styled/inputComment.ts
