+++ FE/issuetracker/src/components/Issue/styled/issueLabel.ts
