+++ FE/issuetracker/src/components/Issue/styled/issueTitleContent.js
